requirements.txt
ultralytics>=8.0.0
opencv-python>=4.5.0
numpy>=1.20.0
mysql-connector-python>=8.0.0